test: move logging config to separate package
This will allow us to read config variables in ginkgo-ext package before
ginkgo creates a test tree.
