+++ test/logger/logs.go
@@ -1,4 +1,4 @@
-package config
+package logger
 
 import (
 	"bytes"
